[feat] data 생성
